refactor: deleted some unecessary files and moved  into the utils folder
