--- artists.csv
@@ -1,4 +1,0 @@
-1,ABBA,1992
-2,Roxette,1986
-3,Europe,1979
-4,The Cardigans,1992
