--- edges.csv
@@ -1,91 +1,0 @@
-:START_ID,:END_ID,:TYPE,strength,description
-w1,l1,CONTAINS,,World contains this location
-w1,l2,CONTAINS,,World contains this location
-w1,l3,CONTAINS,,World contains this location
-w1,c1,CONTAINS,,World contains this character
-w1,c2,CONTAINS,,World contains this character
-w1,c3,CONTAINS,,World contains this character
-w1,e1,CONTAINS,,World contains this event
-w1,e2,CONTAINS,,World contains this event
-w1,e3,CONTAINS,,World contains this event
-w2,l4,CONTAINS,,World contains this location
-w2,l5,CONTAINS,,World contains this location
-w2,l6,CONTAINS,,World contains this location
-w2,c4,CONTAINS,,World contains this character
-w2,c5,CONTAINS,,World contains this character
-w2,c6,CONTAINS,,World contains this character
-w2,e4,CONTAINS,,World contains this event
-w2,e5,CONTAINS,,World contains this event
-w2,e6,CONTAINS,,World contains this event
-w3,l7,CONTAINS,,World contains this location
-w3,l8,CONTAINS,,World contains this location
-w3,l9,CONTAINS,,World contains this location
-w3,c7,CONTAINS,,World contains this character
-w3,c8,CONTAINS,,World contains this character
-w3,c9,CONTAINS,,World contains this character
-w3,e7,CONTAINS,,World contains this event
-w3,e8,CONTAINS,,World contains this event
-w3,e9,CONTAINS,,World contains this event
-c1,c2,MOTHER_OF,,Lyralei is mother of Thornwick
-c2,c1,SON_OF,,Thornwick is son of Lyralei
-c2,c3,FATHER_OF,,Thornwick is father of Dewdrop
-c3,c2,DAUGHTER_OF,,Dewdrop is daughter of Thornwick
-c1,c3,GRANDMOTHER_OF,,Lyralei is grandmother of Dewdrop
-c3,c1,GRANDDAUGHTER_OF,,Dewdrop is granddaughter of Lyralei
-c5,c4,CREATOR_OF,,CORE-3 created ARIA-7
-c4,c5,CREATED_BY,,ARIA-7 was created by CORE-3
-c5,c6,CREATOR_OF,,CORE-3 created NOVA-1
-c6,c5,CREATED_BY,,NOVA-1 was created by CORE-3
-c4,c6,SIBLING_OF,,ARIA-7 and NOVA-1 are siblings
-c6,c4,SIBLING_OF,,NOVA-1 and ARIA-7 are siblings
-c7,c8,PARENT_OF,,Rosalind is parent of Jasper
-c8,c7,OFFSPRING_OF,,Jasper is offspring of Rosalind
-c8,c9,PARENT_OF,,Jasper is parent of Lily
-c9,c8,OFFSPRING_OF,,Lily is offspring of Jasper
-c7,c9,GRANDPARENT_OF,,Rosalind is grandparent of Lily
-c9,c7,GRANDCHILD_OF,,Lily is grandchild of Rosalind
-e1,l1,HAPPENED_AT,,The Great Withering occurred at Crystal Clearing
-e2,l2,HAPPENED_AT,,The Healing Ritual took place at Moonbeam Bridge
-e3,l3,HAPPENED_AT,,The Celebration of Renewal was held at Starlight Grove
-e4,l4,HAPPENED_AT,,The Awakening Protocol occurred at Central Processing Plaza
-e5,l5,HAPPENED_AT,,The Knowledge Transfer took place at Memory Archives
-e6,l6,HAPPENED_AT,,The Unity Ceremony was held at Harmony Gardens
-e7,l7,HAPPENED_AT,,The Drought Warning was given at The Great Meadow
-e8,l8,HAPPENED_AT,,The Rain Summoning occurred at Pollen Parliament
-e9,l9,HAPPENED_AT,,The Bloom Festival took place at Sunbeam Sanctuary
-c1,e1,PRESENT_AT,,Lyralei witnessed the Great Withering
-c2,e1,PRESENT_AT,,Thornwick witnessed the Great Withering
-c1,e2,PRESENT_AT,,Lyralei participated in the Healing Ritual
-c2,e2,PRESENT_AT,,Thornwick participated in the Healing Ritual
-c3,e2,PRESENT_AT,,Dewdrop participated in the Healing Ritual
-c1,e3,PRESENT_AT,,Lyralei attended the Celebration of Renewal
-c2,e3,PRESENT_AT,,Thornwick attended the Celebration of Renewal
-c3,e3,PRESENT_AT,,Dewdrop attended the Celebration of Renewal
-c4,e4,PRESENT_AT,,ARIA-7 was the subject of the Awakening Protocol
-c5,e4,PRESENT_AT,,CORE-3 conducted the Awakening Protocol
-c4,e5,PRESENT_AT,,ARIA-7 received knowledge in the Knowledge Transfer
-c5,e5,PRESENT_AT,,CORE-3 shared knowledge in the Knowledge Transfer
-c6,e5,PRESENT_AT,,NOVA-1 observed the Knowledge Transfer
-c4,e6,PRESENT_AT,,ARIA-7 participated in the Unity Ceremony
-c5,e6,PRESENT_AT,,CORE-3 participated in the Unity Ceremony
-c6,e6,PRESENT_AT,,NOVA-1 participated in the Unity Ceremony
-c7,e7,PRESENT_AT,,Rosalind issued the Drought Warning
-c8,e7,PRESENT_AT,,Jasper heard the Drought Warning
-c7,e8,PRESENT_AT,,Rosalind participated in the Rain Summoning
-c8,e8,PRESENT_AT,,Jasper participated in the Rain Summoning
-c9,e8,PRESENT_AT,,Lily participated in the Rain Summoning
-c7,e9,PRESENT_AT,,Rosalind celebrated at the Bloom Festival
-c8,e9,PRESENT_AT,,Jasper celebrated at the Bloom Festival
-c9,e9,PRESENT_AT,,Lily celebrated at the Bloom Festival
-e1,e2,HAPPENED_BEFORE,,The Great Withering happened before the Healing Ritual
-e2,e1,HAPPENED_AFTER,,The Healing Ritual happened after the Great Withering
-e2,e3,HAPPENED_BEFORE,,The Healing Ritual happened before the Celebration of Renewal
-e3,e2,HAPPENED_AFTER,,The Celebration of Renewal happened after the Healing Ritual
-e4,e5,HAPPENED_BEFORE,,The Awakening Protocol happened before the Knowledge Transfer
-e5,e4,HAPPENED_AFTER,,The Knowledge Transfer happened after the Awakening Protocol
-e5,e6,HAPPENED_BEFORE,,The Knowledge Transfer happened before the Unity Ceremony
-e6,e5,HAPPENED_AFTER,,The Unity Ceremony happened after the Knowledge Transfer
-e7,e8,HAPPENED_BEFORE,,The Drought Warning happened before the Rain Summoning
-e8,e7,HAPPENED_AFTER,,The Rain Summoning happened after the Drought Warning
-e8,e9,HAPPENED_BEFORE,,The Rain Summoning happened before the Bloom Festival
-e9,e8,HAPPENED_AFTER,,The Bloom Festival happened after the Rain Summoning
