--- nodes.csv
@@ -1,31 +1,0 @@
-nodeId:ID,name,description,userId,createdAt,:LABEL
-w1,Luminwood Realm,A mystical forest realm where fairy folk dwell among glowing trees and crystal streams,user1,2024-01-01,World
-w2,Nexus Prime,A gleaming metropolis where advanced robots have built a utopian society,user2,2024-01-02,World
-w3,The Verdant Expanse,A peaceful meadow world ruled by wise and benevolent flower beings,user3,2024-01-03,World
-l1,Crystal Clearing,A shimmering meadow where fairy councils gather beneath ancient oak trees,user1,2024-01-01,Location
-l2,Moonbeam Bridge,An ethereal bridge of solidified moonlight spanning a rushing river,user1,2024-01-01,Location
-l3,Starlight Grove,A sacred grove where the most powerful fairy magic is practiced,user1,2024-01-01,Location
-l4,Central Processing Plaza,The heart of the robot city where important announcements are made,user2,2024-01-02,Location
-l5,Memory Archives,A vast library containing the collective knowledge of all robot-kind,user2,2024-01-02,Location
-l6,Harmony Gardens,Beautiful geometric gardens where robots gather for contemplation,user2,2024-01-02,Location
-l7,The Great Meadow,An endless field of colorful flowers swaying in gentle breezes,user3,2024-01-03,Location
-l8,Pollen Parliament,A circular arrangement of the tallest flowers where decisions are made,user3,2024-01-03,Location
-l9,Sunbeam Sanctuary,A quiet grove where flowers go to absorb sunlight and meditate,user3,2024-01-03,Location
-c1,Lyralei the Wise,An ancient fairy queen with gossamer wings and silver hair,user1,2024-01-01,Character
-c2,Thornwick,Lyralei's son with emerald wings and a talent for plant magic,user1,2024-01-01,Character
-c3,Dewdrop,Thornwick's daughter with rainbow wings and the power to heal,user1,2024-01-01,Character
-c4,ARIA-7,A sophisticated AI with a chrome chassis and advanced learning protocols,user2,2024-01-02,Character
-c5,CORE-3,ARIA-7's creator and mentor with a bulky protective frame,user2,2024-01-02,Character
-c6,NOVA-1,CORE-3's first creation with sleek design and incredible processing speed,user2,2024-01-02,Character
-c7,Rosalind the Magnificent,A towering sunflower with petals that shimmer like gold,user3,2024-01-03,Character
-c8,Jasper the Gentle,Rosalind's offspring with deep purple petals and soothing fragrance,user3,2024-01-03,Character
-c9,Lily the Bright,Jasper's child with pure white petals and the ability to purify water,user3,2024-01-03,Character
-e1,The Great Withering,A mysterious blight threatens the fairy realm's ancient trees,user1,2024-01-01,Event
-e2,The Healing Ritual,The fairy family performs a powerful ritual to restore the forest,user1,2024-01-01,Event
-e3,The Celebration of Renewal,The realm celebrates the forest's recovery with a grand festival,user1,2024-01-01,Event
-e4,The Awakening Protocol,ARIA-7 is first activated and begins learning about the world,user2,2024-01-02,Event
-e5,The Knowledge Transfer,CORE-3 shares decades of accumulated wisdom with ARIA-7,user2,2024-01-02,Event
-e6,The Unity Ceremony,All three robot generations gather to celebrate their bond,user2,2024-01-02,Event
-e7,The Drought Warning,The flowers detect an approaching drought that threatens all life,user3,2024-01-03,Event
-e8,The Rain Summoning,The flower family combines their powers to call life-giving rain,user3,2024-01-03,Event
-e9,The Bloom Festival,The entire meadow erupts in magnificent flowers celebrating the rain,user3,2024-01-03,Event
