+++ src/routes/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import driver from '$lib/db/neo4j';
 	import { Integer, Neo4jError, Node, Session } from 'neo4j-driver';
-	import seedData from '$lib/db/create-utils';
+	import seedData from '$lib/utils/create-utils';
 
 	interface UserProps {
 		name: string;
